:art: Separate context and compilation
