+++ packages/cli/src/index.js
@@ -156,7 +156,7 @@ command
           sourceMap: config.output.sourceMap,
           sourceMapPath: config.output.sourceMapPath,
         }).then(async function(generated) {
-          const outputDirectory = Path.resolve(pundle.config.compilation.rootDirectory, config.output.rootDirectory)
+          const outputDirectory = Path.resolve(pundle.config.context.rootDirectory, config.output.rootDirectory)
           const outputFilePath = Path.resolve(outputDirectory, config.output.bundlePath)
           const outputSourceMapPath = Path.resolve(outputDirectory, config.output.sourceMapPath)
           FS.writeFileSync(outputFilePath, generated.contents)
@@ -170,7 +170,7 @@ command
       }
       return promise.catch(function(error) {
         process.exitCode = 1
-        pundle.compilation.report(error)
+        pundle.context.report(error)
       })
     }).catch(function(error) {
       process.exitCode = 1
