--- packages/pundle/src/compilation/types.js
@@ -1,8 +1,0 @@
-/* @flow */
-
-import type { ComponentAny } from 'pundle-api/types'
-
-export type ComponentEntry = {
-  config: Object,
-  component: ComponentAny,
-}
